requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
amqp==5.0.1               # via kombu
asgiref==3.2.10           # via django
billiard==3.6.3.0         # via celery
cachecontrol==0.12.6      # via firebase-admin
cachetools==4.1.1         # via google-auth
celery==5.0.0             # via -r requirements.in
certifi==2020.6.20        # via requests
cffi==1.14.3              # via google-crc32c
chardet==3.0.4            # via requests
click-didyoumean==0.0.3   # via celery
click-repl==0.1.6         # via celery
click==7.1.2              # via celery, click-didyoumean, click-repl
coreapi==2.3.3            # via drf-yasg
coreschema==0.0.4         # via coreapi, drf-yasg
coverage==5.3             # via -r requirements.in
django-extensions==3.0.9  # via -r requirements.in
django-multiselectfield==0.1.12  # via -r requirements.in
django==3.1.2             # via -r requirements.in, django-multiselectfield, djangorestframework, drf-yasg
djangorestframework==3.11.1  # via -r requirements.in, drf-yasg
drf-yasg==1.17.1          # via -r requirements.in
firebase-admin==4.4.0     # via -r requirements.in
google-api-core[grpc]==1.22.4  # via firebase-admin, google-api-python-client, google-cloud-core, google-cloud-firestore
google-api-python-client==1.12.3  # via firebase-admin
google-auth-httplib2==0.0.4  # via google-api-python-client
google-auth==1.22.1       # via google-api-core, google-api-python-client, google-auth-httplib2, google-cloud-storage
google-cloud-core==1.4.3  # via google-cloud-firestore, google-cloud-storage
google-cloud-firestore==1.9.0  # via firebase-admin
google-cloud-storage==1.31.2  # via firebase-admin
google-crc32c==1.0.0      # via google-resumable-media
google-resumable-media==1.1.0  # via google-cloud-storage
googleapis-common-protos==1.52.0  # via google-api-core
grpcio==1.32.0            # via google-api-core
gunicorn==20.0.4          # via -r requirements.in
httplib2==0.18.1          # via google-api-python-client, google-auth-httplib2
idna==2.10                # via requests
inflection==0.5.1         # via drf-yasg
itypes==1.2.0             # via coreapi
jinja2==2.11.2            # via coreschema
kombu==5.0.2              # via celery
markupsafe==1.1.1         # via jinja2
msgpack==1.0.0            # via cachecontrol
packaging==20.4           # via drf-yasg
pillow==7.2.0             # via -r requirements.in
prompt-toolkit==3.0.7     # via click-repl
protobuf==3.13.0          # via google-api-core, googleapis-common-protos
psycopg2-binary==2.8.6    # via -r requirements.in
psycopg2==2.8.6           # via -r requirements.in
pyasn1-modules==0.2.8     # via google-auth
pyasn1==0.4.8             # via pyasn1-modules, rsa
pycparser==2.20           # via cffi
pydotplus==2.0.2          # via -r requirements.in
pyfcm==1.4.7              # via -r requirements.in
pyparsing==2.4.7          # via packaging, pydotplus
python-decouple==3.3      # via -r requirements.in
pytz==2020.1              # via celery, django, google-api-core, google-cloud-firestore
pyyaml==5.3.1             # via -r requirements.in
requests==2.24.0          # via cachecontrol, coreapi, google-api-core, google-cloud-storage, pyfcm
rsa==4.6                  # via google-auth
ruamel.yaml.clib==0.2.2   # via ruamel.yaml
ruamel.yaml==0.16.12      # via drf-yasg
six==1.15.0               # via click-repl, drf-yasg, google-api-core, google-api-python-client, google-auth, google-auth-httplib2, google-resumable-media, grpcio, packaging, protobuf
sqlparse==0.4.0           # via django
uritemplate==3.0.1        # via coreapi, drf-yasg, google-api-python-client
urllib3==1.25.10          # via requests
vine==5.0.0               # via amqp, celery
wcwidth==0.2.5            # via prompt-toolkit

# The following packages are considered to be unsafe in a requirements file:
# setuptools
